Add `observed_at' column to track last seen of a stream
With this commit twiddler now tracks the last time it has seen the stream
(every time the list of streams are queried, their corresponding
`observed_at' column is updated with current time), which allows for a
proper restart handling, while before the stream would still be reported
if the restart happens an hour after the stream has started.

To test this behaviour, a `clock' package was added that allows to mock
time source in tests with a fixed, adjustable time.

Along with these, some refactoring went into this commit, such as `live'
field of `Tracker' is now protected with a mutex, remains of discord-related
code was moved into messenger/discord, user cache can now expire
(currently expiration fixed at 48 hours).